ft state up by moving List<Task> tasks into tasks_screen.dart and add a propert listOfTasks for AddTasksScreen()
Lifted the state up by moving tasks into tasks_screen.dart, though I
have still left a copy of tasks in task_list in order to prevent any
errors for now. In order to implement add functionality I am trying to
pass list of Task() to AddTaskScreen() so that I can modify it in
AddTaskScreen() on tapping Add button. Add functionality is yet to be
completed. For now I have checked that listOfTasks is initially empty
and elements are being added to it when I tap Add button. But the values
in list which are passed to AddTasksScreen() aren't getting stored into
listOfTasks in AddTaskScreen().